dcim_backend_fastapi/app/requirements.txt
############################################################
# Core Framework & ASGI stack
############################################################
fastapi==0.115.2
starlette==0.40.0           # middleware + streaming responses
uvicorn[standard]==0.30.1

############################################################
# Database & ORM
############################################################
SQLAlchemy==2.0.34
oracledb==3.3.0             # Oracle DB driver (recommended)
# cx_Oracle==8.3.0            # (Alternative driver – optional)
alembic==1.13.2


############################################################
# Auth / Security
############################################################
bcrypt==4.1.2                # for frontend-compatible password hashing
cryptography==43.0.1         # (optional: for AES/RSA decryption later)
PyJWT==2.8.0                 # JWT token decoding for logging
ldap3==2.9.1                 # LDAP/Active Directory authentication

############################################################
# Pydantic / Validation
############################################################
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.3.0
python-multipart==0.0.20

############################################################
# Utility
############################################################
python-dotenv==1.0.1         # load .env files
typing_extensions==4.12.2
openpyxl==3.1.5              # Excel/XLXS helpers for templates
pandas==2.2.2                # CSV parsing for bulk upload + exports
python-json-logger==2.0.7    # Structured JSON logging formatter


############################################################
# Testing (optional but recommended)
############################################################
pytest==8.3.1
pytest-asyncio==0.23.7
httpx==0.27.0                # test FastAPI endpoints